extract embedding modules from cluster/ to embeddings/luxical/
Splits the Luxical-One embed pipeline out of the clustering subtree so
the embedding step (Zephyr map-only emit of int8 EmbeddingAttrData) is
discoverable on its own and can be reused by non-clustering callers.

Moved: embed_source.py, bench_batch_size.py, exp_fixup_oom.py.
Updated imports in assign.py, sample.py, exp_full_clusters.py, exp_smoke.py.
