+++ experiments/datakit/cluster/exp_smoke.py
@@ -38,10 +38,14 @@ from marin.execution.step_spec import StepSpec  # noqa: E402
 from rigging.filesystem import marin_temp_bucket  # noqa: E402
 
 from experiments.datakit.cluster.assign import assign_source  # noqa: E402
-from experiments.datakit.cluster.embed_source import LUXICAL_REPO, LUXICAL_WEIGHTS_FILE, embed_source  # noqa: E402
 from experiments.datakit.cluster.sample import sample_centroid_inputs  # noqa: E402
 from experiments.datakit.cluster.summarize import summarize_at_k  # noqa: E402
 from experiments.datakit.cluster.train import train_centroids  # noqa: E402
+from experiments.datakit.embeddings.luxical.embed_source import (  # noqa: E402
+    LUXICAL_REPO,
+    LUXICAL_WEIGHTS_FILE,
+    embed_source,
+)
 
 logger = logging.getLogger(__name__)
